Enable adaptive transparency for Plasma 5.22+
This makes the option available in the panel settings, so users
will also be able to disable transparency in addition to making
it adaptive. (Adaptive just means that transparency is disabled
when an app is maximised, this applies to both the panel and
popups coming from it, including the application launcher).